ref(metrics): Extract MRISelect as separate component
